Perbaiki migrasi bisa dijalankan berulang-ulang

+++ donjo-app/models/migrations/Migrasi_2004_ke_2005.php
@@ -20,7 +20,8 @@ class Migrasi_2004_ke_2005 extends CI_model {
 		$this->db->where('id_aset', 3423)->update('tweb_aset', array('nama' => 'JALAN'));
 		$this->db->where('id', 79)->update('setting_modul', array('url'=>'api_inventaris_kontruksi', 'aktif'=>'1'));
 		// Hapus field urut di tabel artikel krn tdk dibutuhkan
-		$this->db->query('ALTER TABLE `artikel` DROP COLUMN `urut`');
+		if ($this->db->field_exists('urut', 'artikel'))
+			$this->db->query('ALTER TABLE `artikel` DROP COLUMN `urut`');
 	}
 	
 }
